Consolidate OpenAI session_id/agent_version propagation tests into one file
