+++ dd-java-agent/instrumentation/openai-java/openai-java-3.0/src/test/java/datadog/trace/instrumentation/openai_java/LlmObsContextPropagationForkedTest.java
@@ -25,14 +25,17 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
 /**
- * Verifies that auto-instrumented openai.request spans inherit agent_version from an active LLMObs
- * parent context (e.g. a manual agent span), and that a stale context left over from an unrelated
- * trace does not leak session_id or agent_version onto the span. Forked + @WithConfig used together
- * so the LLMObs system property is in place before the agent installs and there's no leakage from
- * prior test state.
+ * Verifies that auto-instrumented openai.request spans inherit session_id and agent_version from an
+ * active LLMObs parent context, and that a stale context left over from an unrelated trace does not
+ * leak either tag onto the span. Forked + @WithConfig used together so the LLMObs system property
+ * is in place before the agent installs and there's no leakage from prior test state.
+ *
+ * <p>The mock OpenAI backend returns a minimal 200 response — the test asserts on the span tag set
+ * by OpenAiDecorator.afterStart(), which runs before the HTTP response is parsed, so the response
+ * body shape doesn't matter for what's being tested.
  */
 @WithConfig(key = "llmobs.enabled", value = "true")
-class AgentVersionPropagationForkedTest extends AbstractInstrumentationTest {
+class LlmObsContextPropagationForkedTest extends AbstractInstrumentationTest {
 
   private static HttpServer mockServer;
   private static OpenAIClient openAiClient;
@@ -67,6 +70,46 @@ class AgentVersionPropagationForkedTest extends AbstractInstrumentationTest {
       mockServer = null;
     }
     openAiClient = null;
+  }
+
+  @Test
+  void openAiRequestSpanInheritsSessionIdFromActiveContext() throws Exception {
+    String expectedSessionId = "session-propagation-test-abc";
+
+    AgentSpan parentSpan = AgentTracer.startSpan("test", "parent");
+    try (ContextScope ignored1 = AgentTracer.activateSpan(parentSpan)) {
+      try (ContextScope ignored2 =
+          LLMObsContext.attach(parentSpan.spanContext(), expectedSessionId)) {
+        try {
+          openAiClient.chat().completions().create(buildMinimalChatParams());
+        } catch (Exception ignored) {
+          // Mock server returns no body — the SDK may throw on parse. The span we care about
+          // is already created by the instrumentation advice before this point.
+        }
+      }
+    } finally {
+      parentSpan.finish();
+    }
+
+    writer.waitForTraces(1);
+    DDSpan openAiSpan = findSpanByOperationName(writer, "openai.request");
+    assertNotNull(openAiSpan, "openai.request span should have been created");
+    assertEquals(expectedSessionId, openAiSpan.getTag("_ml_obs_tag.session_id"));
+  }
+
+  @Test
+  void openAiRequestSpanHasNoSessionIdWhenNoLlmObsContext() throws Exception {
+    try {
+      openAiClient.chat().completions().create(buildMinimalChatParams());
+    } catch (Exception ignored) {
+      // Mock server returns no body — the SDK may throw on parse. The span we care about
+      // is already created by the instrumentation advice before this point.
+    }
+
+    writer.waitForTraces(1);
+    DDSpan openAiSpan = findSpanByOperationName(writer, "openai.request");
+    assertNotNull(openAiSpan, "openai.request span should have been created");
+    assertNull(openAiSpan.getTag("_ml_obs_tag.session_id"));
   }
 
   @Test
